UserService, UserController

+++ src/main/java/dev/projekt/students/courses/user/UserService.java
@@ -1,8 +1,12 @@
 package dev.projekt.students.courses.user;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class UserService {
@@ -11,6 +15,9 @@ public class UserService {
     private UserRepository userRepository;
 
     public User createUser(User user) {
+        // TODO szyfrowanie hasla (Spring Security -> pom.xml)
+        // String encryptedPassword = passwordEncoder.encode(user.getPassword());
+        // user.setPassword(encryptedPassword);
         return userRepository.save(user);
     }
 
@@ -18,11 +25,14 @@ public class UserService {
         return userRepository.findAll();
     }
 
-    public User getUserById(Long id) {
-        return userRepository.findById(id).orElse(null);
+    public Optional<User> getUserById(Integer id) {
+        return userRepository.findById(id);
     }
 
-    public void deleteUser(Long id) {
+    public void deleteUser(Integer id) {
+        if (!userRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User not found.");
+        }
         userRepository.deleteById(id);
     }
 }
